We need to clarify these copyright issues...
In France, no mention of copyright is needed: original work is
protected from the moment of creation. No mention of copyright
is needed on web sites and most of them seem to be omitting it
nowadays (especially CEA sites).
In the US, you may take this one step further: copyright registration.
In any case, I believe mentionning the copyright is not really needed;
it is more of a deterrent.
If you do mention the copyright, at least refer to a legal entity.