Copyright
I don't think the real intent was to mention copyright here.
Instead you might want to explain the web site is based on
PIWS, which is developped by the Neurospin platform based
on the brainomics cubes. Why not add a special page for that?

About copyright:
In France, no mention of copyright is needed: original work
is protected from the moment of creation. No mention of copyright
is needed on web sites and most of them seem to be omitting it
nowadays (especially CEA sites).
In the US, you may indeed take this one step further: copyright
registration.

In any case, I believe mentionning the copyright is not really
needed; it is more of a deterrent. Let's keep the copyright for
now - but at least refer to a legal entity.