This hack only caused crashes
Both in debug mode (cubicweb-ctl start -D) and release mode
(cubicweb-ctl start) this piece of code would result in crashes
in the docutils module.

I do not see how this piece of code could have avoided crashes:
it changes the current directory which may precisely trigger the
bug of docutils where it depends on relative paths (that depend
on the current dir) instead of using absolute paths.

The only way this piece of code could help is if the initial current
directory is precisely VIRTUAL_ENV - which is usually the case.

In any case, if we do reproduce the issue where the current directory
is reset behind the back of docutils, this piece of code will not be
the proper solution: it depends on running cubicweb-ctl from a specific
directory. Instead find where os.chdir() is called elsewhere and
address the problem exactly there.